wrappers/services.go: add `after` svcs to Requires, inverse for `before`
For daemons in a snap, we need to make the dependencies "hard" using systemd's "Requires=", directive. This is because when the snap is installed, all of the services are started simultaneously with one transaction a la `systemctl start <all services>`, and unless the dependencies are "hard" using Requires, systemctl when starting these services ignores the "After"/"Before" directives. It always obeys "Requires" however, so to ensure reliable operation of using "after"/"before" during the install task, so we add dependencies using that as well as the "After"/"Before" dependencies.