cmd/snap: use Name not Full() in message; improve tests
In the deprecation message we were using Full() instead Name, which
meant "/stable" got printed as "latest/stable" even though what we
want (and use) is actually "stable". This fixes that.

Also improved test by checking that the client sends what we want, and
that ///foo/stable// is also covered just in case.

Thanks to @pedronis, @mvo5 and @anonymouse64 for spotting these.