osutil/strace/timing.go: also handle SIGKILL for exec's
We already handle SIGCHLD for a child process dying and SIGTERM for normal
process exits, but we are missing SIGKILL right now which shows up under a
different format. So we should catch those and mark processes that die from
SIGKILL as finished too so they show up in the output.

Also add a doc comment to TraceExecveTimings and fix a typo in a comment.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>