modeenv: add current_kernels for list of trusted kernels
This will be mutated by the boot package as we add kernel snaps, and provides us
more enforcement of what kernel snaps are safe to use in the initramfs when we
need to mount the kernel snap to provide kernel modules.

Note that since the config parser only supports scalars, we need to store the
list of kernels as a comma delimited string. There are tests to ensure this
encoding is robust against garbage in the config file.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>