boot: set CurrentKernels in modeenv when we call setNext for a kernel
This change adds basic support for adding kernels to the list of trusted kernels
in the modeenv, which `snap-bootstrap initramfs-mounts` will use to verify that
the snap referenced by the kernel.efi symlink is a valid kernel snap to mount
modules from for booting.

Many more tests now need to write/read the modeenv, so also update tests for
that.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>