bootstate20: always re-write current_kernels manually
Rather than appending to it, we can be more robust by saving the current kernel
and writing current_kernels as the current fallback kernel and the try kernel
exactly.

This enforces the fact that current_kernels[0] is our fallback kernel and
current_kernels[1] is the try kernel always, and also protects against an attack
where the attacker swaps the values of the kernel.efi and try-kernel.efi
symlinks on ubuntu-boot, as the initramfs can now rely on the current_kernels to
know exactly which kernel we expected to be the try kernel and the fallback
kernel.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>