tests/interfaces-audio-playback-record: port pulseaudio test changes
These changes make it easier to debug what's going on when the test fails, and
simplifies the test a bit using the as_user helper. Ideally this should
eventually switch to using session-tool to run things as the user, as pulseaudio
is likely doing some kind of user session interaction which doesn't work well.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>