secboot,cmd/snap-bootstrap: refactor device handling
secboot.UnlockResult now contains a DecryptedDevice and a Device, where the
DecryptedDevice is set to the decrypted mapper device if decryption is
successful, and Device is always set to the encrypted block device if it is
found - regardless of whether it was successfully decrypted or not.

This allows us to implement more correct state handling in the initramfs, where
we can better distinguish between the case of "found but not unlocked" and "not
found and not unlocked".

Also update the exported Device field of degraded.json to have the following
meaning:
- Device is set to "" if the encrypted device and the possible unencrypted
  device are not found.
- Device is set to the block device of the encrypted device if the encrypted
  block device is found but not successfully decrypted.
- Device is set to the decrypted mapper device if decryption is successful.
- Device is set to the block device of the unencrypted device if that was found.

This also contains an internal refactoring of the recoverDegradedState struct to
now track separately the "rawDevice" and "decryptRawDevice" which have the
following meanings:
- rawDevice is always set to the block device of the partition in question if
  the partition in question can be found
- decryptRawDevice is set to the block device of the encrypted partition if it
  has not yet been decrypted (thus also if it failed to be decrypted)
- decryptRawDevice is set to the decrypted mapper device if it was successfully
  decrypted

Additionally, we also now don't have a "locate unencrypted save" state, and that
step is now also implemented inside unlockSaveFallbackKey() as well since the
secboot function we use there is good for both encrypted and unencrypted
ubuntu-save's.

Finally, add a new unit test for the unencrypted case where data is absent from
the disk, and update the degraded mode test names to be more clear what they are
supposed to be testing.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>