[WIP] strutil: add KernelCommandLineKeyValuePairs
Ideally this would be in boot, along-side the other kernel command line parsing
functions, but unfortunately we need to use it from bootloader, which boot
imports, and so we would have an import cycle if we did this.

Instead just stash it in strutil for now and fix up where this should actually
live later.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>