bootloader/lk.go: adjust partition locating to use secure kernel cmdline
We only want to use partition names from the disk that the bootloader says is
trusted, so what we do is we parse the kernel command line (which will be set by
the bootloader) for the kernel command line parameter "snapd_lk_boot_disk", and
use that to find the partitions with the relevant names for the structures.

This is all a bit of a hack, since we really need to be able to return errors
from envFile(), but it's used a bit in places that don't return errors, and we
can't really do anything if any of the steps fail, and it's entirely possible
that something could actually fail, so just panic for now. We should probably
adjust things so that errors can be returned here properly.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>