cmd/snap-confine/snap-confine.c: clear LD_PRELOAD for snaps
The value of LD_PRELOAD is used to load additional libraries for applications
which are dynamically linked, but the issue with this is that the setting of
LD_PRELOAD for a strict snap will not make sense because typical libraries that
this environment variable would be included here will not be visible to a strict
snap. As such, clear the value of the environment variable in snap-confine.

Note that snaps can still use LD_PRELOAD if they define it in their wrapper
scripts or in the environment section of their snap{,craft}.yaml, because the
values specified there will be applied after snap-confine runs during snap-exec.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>