o/storecontext/context.go: add support for env vars to override cloud bits
This lets one set the cloud specific bits on a device that might not have had
cloud-init run, but is still located in a region that wants to send specific
cloud information to the store so that the store can route the CDN, etc. in the
response for downloading snaps to a regional location.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>